requirements.txt
av==12.3.0
numpy==2.1.1
opencv_contrib_python==4.5.3.56
opencv_python==4.9.0.80
PyYAML==6.0
PyYAML==6.0.2
st_pages==1.0.1
streamlit==1.12.0
streamlit_authenticator==0.2.2
streamlit_webrtc==0.47.9
